% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clcc.R
\name{clcc}
\alias{clcc}
\title{Calculate the CLCC indicator}
\usage{
clcc(
  data_path,
  use_weights = FALSE,
  weights_path = NULL,
  func_unit = "km",
  label_digits = 3,
  plot_phases = FALSE,
  plot_phases_critical = FALSE,
  plot_phases_critical_version = "EU",
  price_source = "latest"
)
}
\arguments{
\item{data_path}{A character vector. Path to the folder in which raw xlsx files are stored.}

\item{use_weights}{A logical value. If set to `TRUE`, the function uses the critical weights}

\item{weights_path}{A character vector. Path to the file containing the critical weights for each commodity and phase.}

\item{func_unit}{A string, the functional unit of the LCA analysis the inventories refer to. This parameter
is needed to plot the data it is not used to compute the indicator.}

\item{label_digits}{An integer: the number of digits to display in the plots.}

\item{plot_phases}{A logical value. if set to `TRUE` the function returns a plot object (`ggplot2`) displaying
the relative importance of each life cycle phase on the CLCC indicator.}

\item{plot_phases_critical}{A logical value. if set to `TRUE` and `plot_phases` is also set to `TRUE`, the
plot returned refers to the Critical CLCC indicator.}

\item{plot_phases_critical_version}{A string. If set to `EU`, the critical CLCC indicator is based on the list of critical
materials of the European Union. If it is set to `IEA` the International Energy Agency list is used instead. Default is set to `EU`}

\item{price_source}{A string. If set to `latest` latest price data are used. If set to `previous`,
the previous' year data are used instead.}
}
\value{
A list of 2 elements: a tibble containing the CLCC indicator calculated for each object and phase
    and a `ggplot` object containing the plot of the results.
}
\description{
This function calculates the CLCC indicator, a measure of resource scarcity,
for a product/process during its life cycle. The function automatically loads
the latest prices available and SimaPro inventories using other functions of
the clccr package.
The function can also return the CLCC indicator calculated taking into account
critical materials only.
}
\examples{
\dontrun{
data_path <- path_to_inventory_folder

clcc(data_path = data_path)

}
}
