% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clcc_mc.R
\name{clcc_mc}
\alias{clcc_mc}
\title{Run Monte Carlo analysis for the CLCC indicator}
\usage{
clcc_mc(
  data_path,
  use_weights = FALSE,
  weights_path = NULL,
  rep = 10000,
  phase = "total",
  critical = FALSE,
  critical_type = "EU",
  func_unit = "km",
  prob_inf_alt = FALSE
)
}
\arguments{
\item{data_path}{A character vector. Path to the folder in which raw xlsx files are stored.}

\item{use_weights}{A logical value. If set to `TRUE`, the function uses the critical weights}

\item{weights_path}{A character vector. Path to the file containing the critical weights for each commodity and phase.}

\item{rep}{Number of Monte Carlo iterations. Default is 10,000}

\item{phase}{LCA phase for which the simulation is run. Default is "total"}

\item{critical}{A logical value. If set to `TRUE`, shares referred to the critical-clcc indicator
are returned. Default is set to `FALSE`.}

\item{critical_type}{A string. If set to `EU`, the critical CLCC indicator is based on the list of critical
materials of the European Union. If it is set to `IEA` the International Energy Agency list is used instead. Default is set to `EU`}

\item{func_unit}{A string, the functional unit of the LCA analysis the inventories refer to. This parameter
is needed to plot the data it is not used to compute the indicator.}

\item{prob_inf_alt}{If TRUE the function computes the probability that an
object's CLCC is lower than that of the other objects belonging to the same project.
Default is set to FALSE.}
}
\value{
If prob_inf_alt is set to FALSE: a nested tibble containing
object names, baseline clcc, all simulations, and the probability that
the indicator is lower than the baseline.
If prob_inf_alt is set to TRUE: a tibble with the probability that each object's
CLCC is lower than that of every other object loaded.
}
\description{
This function loads price and inventory data (using other clccr functions)
and runs a Monte Carlo simulation of the CLCC indicator. A high number
(decided by the user) of random prices are generated using triangular distributions
(parameters of such distributions are mean, min and max prices of the last decade),
which are then used to compute as many values of the CLCC indicator.
The function provides also the probability (computed using the empirical distribution
function) that the CLCC of a given object is lower than the baseline or lower/higher
than that of any other object belonging to the same project.
}
\examples{
\dontrun{

data_path <- path_to_inventory_folder

clcc_mc(data_path = data_path)

clcc_mc(data_path = data_path, rep = 5000, prob_inf_alt = TRUE)

}
}
