% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clcc_prices_ref.R
\docType{data}
\name{clcc_prices_ref}
\alias{clcc_prices_ref}
\title{Dataset with commodity names, codes and reference prices}
\format{
A tibble with variables:
\describe{
\item{comm}{commodity names, as in SimaPro output}
\item{no_comm}{commodity ID}
\item{um}{measurement unit, as in SimaPro output}
\item{source}{Reference price source}
\item{critical}{Whether or not a given commodity belogns to the critical materials group defined by the EU}
\item{code}{Commodity code (depends on the source)}
\item{mean}{Average reference price, expressed in constant euro/kg (m3) at the price levels of the reference year}
\item{min}{Minimum price suring the reference period}
\item{max}{Maximum price suring the reference period}
\item{n_obs}{Number of years over which reference prices are computed}
\item{ref_yr}{Reference year of the analysis}
}
}
\usage{
clcc_prices_ref
}
\description{
A datataset containing all raw materials considered for the computation of the
CLCC indicator and the various sources from which it is possible to gather price
data for each commodity. The dataset also provides information on the preferred
data source for each commodity and on average prices (last 10 years) to be used in
computations. The data set provides also minimum and maximum prices of the reference
period to be used in Monte Carlo analysis
}
\keyword{datasets}
