% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_xlsx_files.R
\name{copy_xlsx_files}
\alias{copy_xlsx_files}
\title{Copies inventories in from an origin to a working folder and replaces extension}
\usage{
copy_xlsx_files(source_folder, destination_folder)
}
\arguments{
\item{source_folder}{Path to the folder in which raw `.XLSX` or `.xlsx` files from SimaPro are stored}

\item{destination_folder}{Path to the folder in which modified `.xlsx` files have to be saved}
}
\description{
This function copies LCA inventories (to be used for CLCC calculations) from an origin to a destination folder. It also
changes the file extension from `.XLSX` to `.xlsx` if needed. Sometimes spreadsheets generated with the LCA software SimaPro
have the extension in caps lock and this can cause the `clcc()` or the `clcc_mc()` functions to crash.
}
\examples{
\dontrun{
source_folder <- path_to_inventory_folder
destination_folder <- path_where_to_save_data

copy_xlsx_files(source_folder, destination_folder)

}
}
