% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical_weights_load_fn.R
\name{critical_weights_load_fn}
\alias{critical_weights_load_fn}
\title{Load critical weights from an Excel file}
\usage{
critical_weights_load_fn(path_weights)
}
\arguments{
\item{path_weights}{A character vector. Path to the file containing the critical weights for each commodity and phase.}
}
\description{
This function leads critical weights from an Excel file. Such weights are project specific. Some materials are considered
critical by the European Commission of the International Energy Agency (IEA) but are not considered as individual flows by
SimaPro. For example the SimaPro flow "Coal, hard" contains "coking coal" which is critical. The weights are therefore needed to
compute the Critical-CLCC indicator.
}
\examples{
\dontrun{
critical_path <- path_to_critical_weights_file

critical_weights_load_fn(path_weights = critical_path)

}
}
