% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exc_rate}
\alias{exc_rate}
\title{USD-EUR Nominal exchange rate}
\format{
A tibble with two variables:
\describe{
\item{year}{year}
\item{exc_rate}{USD per 1 euro}
}
}
\source{
\url{https://db.nomics.world/OECD/MEI}
}
\usage{
exc_rate
}
\description{
Mean annual nominal exchange rate: USD per EUR. Data are retrived from the
Eurostat database. See script in data-raw for details.
}
\keyword{datasets}
