% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdp_defl}
\alias{gdp_defl}
\title{Euro Area GDP deflator}
\format{
A tibble with two variables
\describe{
\item{year}{year}
\item{cur}{Country to which the GDP deflator refers to}
\item{gdp_defl}{GDP deflator}
}
}
\source{
\url{https://db.nomics.world/OECD/MEI}
}
\usage{
gdp_defl
}
\description{
Annual GDP deflator for the Euro Area. Base year 2015.
}
\keyword{datasets}
