% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdp_dfl}
\alias{gdp_dfl}
\title{Euro Area GDP deflator}
\format{
A tibble with two variables
\describe{
\item{year}{year}
\item{gdp_dlf}{EA GDP deflator}
}
}
\source{
\url{https://ec.europa.eu/eurostat/data/database/}
}
\usage{
gdp_dfl
}
\description{
Annual GDP deflator for the Euro Area. Base year 2015.
}
\keyword{datasets}
