% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory_load_fn.R
\name{inventory_load_fn}
\alias{inventory_load_fn}
\title{Load raw inventory data and make them ready for CLCC calculation}
\usage{
inventory_load_fn(data_path)
}
\arguments{
\item{data_path}{A character vector. Path to the folder in which raw xlsx files are stored.
File names must be of the following format: project_name-object_name.xlsx}
}
\value{
A tidy dataset with inventory data organized by project and object
}
\description{
This function loads raw iventory data as xlsx files generated by the LCA SimaPro
software. The function loads the data and organizes them according to the project
they belong. For this reason it is important that xlsx files are named in a
standardized way (see below). The function also converts raw data to measurement
units compatible with those of reference prices.
}
