% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{master_data}
\alias{master_data}
\title{Master file with commodity keys}
\format{
A tibble with variables:
\describe{
\item{comm}{commodity names, as in SimaPro output}
\item{no_comm}{commodity ID}
\item{um}{measurement unit, as in SimaPro output}
\item{comtrade_code}{Comtrade HS 6-digit code}
\item{usitc_code}{USITC 8-digit code}
\item{comext_code}{COMEXT 8-digit code}
\item{ds_code}{Refinitiv Datastream code}
\item{imf_code}{IMF code}
\item{eurostat_code}{Eurostat code}
\item{source}{Reference price source}
\item{critical}{A critical material for the EU, yes/no}
}
}
\usage{
master_data
}
\description{
A datataset containing all raw materials considered for the computation of the
CLCC indicator and the various sources from which it is possible to gather price
data for each commodity. The dataset also provides information on the preferred
data source for each commodity.
}
\keyword{datasets}
