% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{price_comtrade_def}
\alias{price_comtrade_def}
\title{Unit values for some raw materials (Comtrade)}
\format{
A tibble with five variables:
\describe{
\item{year}{year, 10 year period before reference year}
\item{cur}{currency (USD)}
\item{source}{data source}
\item{code}{code, 6-digit HS}
\item{price}{unit value}
}
}
\usage{
price_comtrade_def
}
\description{
A dataset containing unit values (a proxy for market prices) for some raw materials
included in the CLCC indicator. Data are from the UN's Comtrade database, which
provides trade data (6-digit level). Data care downloaded through the UN comtrade API
(see data-raw folder for the script used for downloading the series).
}
\keyword{datasets}
