% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{price_eurostat_def}
\alias{price_eurostat_def}
\title{Electricity prices - European Union average}
\format{
A tibble with five variables:
\describe{
\item{year}{year, 10 year period before reference year}
\item{cur}{currency (EUR)}
\item{source}{data source}
\item{code}{code, eurostat database}
\item{price}{price, excluding taxes and levies}
}
}
\usage{
price_eurostat_def
}
\description{
A dataset containing electricity prices, to be used in the CLCC indicator calculation
as a proxy for the price of electric energy from renewable sources. Data are
from the Eurostat database and refer to the EU and to the price paid by a large
idustrial plant. Data are downloaded through the Eurostat API (see data-raw
folder for the script used for downloading the series).
}
\keyword{datasets}
