% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simapro_codes.R
\docType{data}
\name{simapro_codes}
\alias{simapro_codes}
\title{Dataset with SimaéPro commodity codes}
\format{
A tibble with variables:
\describe{
\item{comp}{LCA compartment, as in SimaPro output}
\item{var1}{LCA variable, as in SimaPro output}
\item{comm}{LCA flow, as in SimaPro output}
\item{code1}{SimaPro code for the commodity, part 1}
\item{code2}{SimaPro code for the commodity, part 2}
\item{formula}{Chemical formula of the commodity, if available}
}
}
\usage{
simapro_codes
}
\description{
Dataset with SimaéPro commodity codes
}
\keyword{datasets}
