% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simapro_export.R
\name{simapro_export}
\alias{simapro_export}
\title{Generate a SimaPro-compatible csv file of characterization factors}
\usage{
simapro_export(groups = FALSE, crit = FALSE)
}
\arguments{
\item{groups}{A logical value (either TRUE or FALSE). If set to TRUE, the commodities are grouped
in macro-categories to make interpretation easier.}

\item{crit}{A logical value that can be selected only when `groups = TRUE`. If it is set to `TRUE` a .csv
file containing critical materials only is generated to be exported to Simapro.}
}
\value{
A SimaPro-compatible csv file
}
\description{
This function generates a csv file with commodity prices (characterization factors) compatible with
SimaPro. Such file can be loaded into SimaPro and used by LCA practitioners to compute the CLCC indicator.
The function automatically asks the user where to save the csv file. This function has no arguments.
}
\examples{
\dontrun{
simapro_export()

simapro_export(critical_only = FALSE)
}
}
