% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/um_converter.R
\name{um_converter}
\alias{um_converter}
\title{Perform measurement unit conversion}
\usage{
um_converter(value, from, to)
}
\arguments{
\item{value}{A numeric value. It can also be a vector. Values to convert.}

\item{from}{A string. The measurement units from which the conversion must be performed.}

\item{to}{A strong. The measurement units to which the conversion must be performed.}
}
\value{
A numeric value (dimensionless): the converted value.
}
\description{
A function that wraps upon `units::set_units` to perform measurement unit conversions,
}
