% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{um_p}
\alias{um_p}
\title{Measurement units of original IMF commodity prices}
\format{
A tibble with three variables:
\describe{
\item{um_from}{original measurement unit}
\item{um_to}{measurement units to which original data mist be converted)}
\item{fct}{conversion factor, not used}
}
}
\usage{
um_p
}
\description{
Primary commoidty prices come in some specific measurement units, which
need to be converted in either kg, m3 or MJ, the measurement units
used for the calculation of the CLCC indicator. The conversion is done using
the *units* package but this dataset provides information on the original
units and those to which original data must be converted.
}
\keyword{datasets}
