% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clcc.R
\name{clcc}
\alias{clcc}
\title{Calculate the CLCC indicator}
\usage{
clcc(path, critical = FALSE)
}
\arguments{
\item{path}{A character vector. Path to the folder in which raw xlsx files are stored.}

\item{critical}{An optional argument. If set to TRUE, the function returns a tibble
containing both the clcc and the critical clcc indicators. Default set to FALSE.}
}
\value{
A tibble containing the CLCC indicator calculated for each object and phase.
}
\description{
This function calculates the CLCC indicator, a measure of resource scarcity,
for a product/process during its life cycle. The function automatically loads
the latest prices available and SimaPro inventories using other functions of
the clccr package.
The function can also return the CLCC indicator calculated taking into account
critical materials only.
}
