% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clcc.R
\name{clcc}
\alias{clcc}
\title{Calculate the CLCC indicator}
\usage{
clcc(
  path,
  func_unit = "km",
  label_digits = 3,
  plot_phases = FALSE,
  plot_phases_critical = FALSE,
  price_source = "2024"
)
}
\arguments{
\item{path}{A character vector. Path to the folder in which raw xlsx files are stored.}

\item{func_unit}{A string, the functional unit of the LCA analysis the inventories refer to. This parameter
is needed to plot the data it is not used to compute the indicator.}

\item{label_digits}{An integer: the number of digits to display in the plots.}

\item{plot_phases}{A logical value. if set to `TRUE` the function returns a plot object (`ggplot2`) displaying
the relative importance of each life cycle phase on the CLCC indicator.}

\item{plot_phases_critical}{A logical value. if set to `TRUE` and `plot_phases` is also set to `TRUE`, the
plot returned refers to the Critical CLCC indicator.}

\item{price_source}{A string. If set to `2023` price sources used in the 2023 RDS report are used. If set to `2024`,
the default, the 2024 updated price sources are used.}
}
\value{
A list of 2 elements: a tibble containing the CLCC indicator calculated for each object and phase
    and a `ggplot` object containing the plot of the results.
}
\description{
This function calculates the CLCC indicator, a measure of resource scarcity,
for a product/process during its life cycle. The function automatically loads
the latest prices available and SimaPro inventories using other functions of
the clccr package.
The function can also return the CLCC indicator calculated taking into account
critical materials only.
}
\examples{
\dontrun{
data_path <- path_to_inventory_folder

clcc(path = data_path)

}
}
