% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clcc_detail.R
\name{clcc_detail}
\alias{clcc_detail}
\title{Compute the realtive share of individual material/energy flows on total CLCC}
\usage{
clcc_detail(
  path,
  critical = FALSE,
  phase_of_int = "total",
  collapse_share = 0.9,
  price_source = "2024"
)
}
\arguments{
\item{path}{A character vector. Path to the folder in which raw xlsx files are stored.}

\item{critical}{A logical value. If set to `TRUE`, shares referred to the critical-clcc indicator
are returned. Default is set to `FALSE`.}

\item{phase_of_int}{A character string. The life-cycle phase for which the relative shares
must be computed. Default is `total`. The user can however select any phase included in the
inventory file.}

\item{collapse_share}{A numeric value. A parameter that the user can use to limit the number of
commodities considered. For example, if `collapse_share = 0.9` (the default) only commodities which
shares sum up to 90% of the total are returned as output. All the others are summed up in the
`others` category. The parameter must be between `0` and `1`.}

\item{price_source}{A string. If set to `2023` price sources used in the 2023 RDS report are used. If set to `2024`,
the default, the 2024 updated price sources are used.}
}
\value{
A list containing a table with the results and a `ggplot` object.
}
\description{
This function computes the relative shares of each commodity considered for the
computation of the CLCC indicator on the total value of the indicator for both the
`baseline` and `critical` versions of the indicator itself.
}
\examples{
\dontrun{
data_path <- path_to_inventory_folder

clcc_detail(path = data_path)

clcc_detail(path = data_path, critical = TRUE, collapse_share = 0.5)

}
}
