% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clcc_detail.R
\name{clcc_detail}
\alias{clcc_detail}
\title{Compute the relative share of individual material/energy flows on total CLCC}
\usage{
clcc_detail(
  data_path,
  use_weights = FALSE,
  weights_path = NULL,
  critical = FALSE,
  critical_type = "EU",
  phase_of_int = "total",
  collapse_share = 0.9,
  price_source = "2024",
  viridis_palette = "viridis",
  palette_direction = 1
)
}
\arguments{
\item{data_path}{A character vector. Path to the folder in which raw xlsx files are stored.}

\item{use_weights}{A logical value. If set to `TRUE`, the function uses the critical weights}

\item{weights_path}{A character vector. Path to the file containing the critical weights for each commodity and phase.}

\item{critical}{A logical value. If set to `TRUE`, shares referred to the critical-clcc indicator
are returned. Default is set to `FALSE`.}

\item{critical_type}{A string. If set to `EU`, the critical CLCC indicator is based on the list of critical
materials of the European Union. If it is set to `IEA` the International Energy Agency list is used instead. Default is set to `EU`}

\item{phase_of_int}{A character string. The life-cycle phase for which the relative shares
must be computed. Default is `total`. The user can however select any phase included in the
inventory file.}

\item{collapse_share}{A numeric value. A parameter that the user can use to limit the number of
commodities considered. For example, if `collapse_share = 0.9` (the default) only commodities which
shares sum up to 90% of the total are returned as output. All the others are summed up in the
`others` category. The parameter must be between `0` and `1`.}

\item{price_source}{A string. If set to `2023` price sources used in the 2023 RDS report are used. If set to `2024`,
the default, the 2024 updated price sources are used.}

\item{viridis_palette}{A string indicating the palette (from the Viridis package to be applied to the plots)}

\item{palette_direction}{A number (either `1` or `-1`) to provide the palette direction}
}
\value{
A list containing a table with the results and a `ggplot` object.
}
\description{
This function computes the relative shares of each commodity considered for the
computation of the CLCC indicator on the total value of the indicator for both the
`baseline` and `critical` versions of the indicator itself.
}
\examples{
\dontrun{
data_path <- path_to_inventory_folder

clcc_detail(data_path = data_path, weights_path = critical_path)

clcc_detail(data_path = data_path,
weights_path = critical_path,
critical = TRUE, collapse_share = 0.5)

}
}
