% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory_load_fn.R
\name{inventory_load_fn}
\alias{inventory_load_fn}
\title{Load raw inventory data and make them ready for CLCC calculation}
\usage{
inventory_load_fn(data_path)
}
\arguments{
\item{data_path}{A character vector. Path to the folder in which raw xlsx files are stored.
File names will be used as object names on which computing the CLCC indicator.}
}
\value{
A tidy dataset with inventory data organized by object
}
\description{
This function loads raw inventory data as xlsx files generated by the LCA SimaPro
software. The function loads all the inventories in a given folder: the file name becomes
the name of the object to be evaluated. The function also converts raw data to measurement
units compatible with those of reference prices.
}
