% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory_load_fn.R
\name{inventory_load_fn}
\alias{inventory_load_fn}
\title{Load raw inventory data and make them ready for CLCC calculation}
\usage{
inventory_load_fn(data_path, use_weights = FALSE, weights_path)
}
\arguments{
\item{data_path}{A character string. Path to the folder in which raw xlsx files are stored.
File names will be used as object names on which computing the CLCC indicator.}

\item{use_weights}{A logical value. If set to `TRUE`, the function uses the critical weights}

\item{weights_path}{A character vector. Path to the file containing the critical weights for each commodity and phase.}
}
\value{
A tidy dataset with inventory data organized by object
}
\description{
This function loads raw inventory data as xlsx files generated by the LCA SimaPro
software. The function loads all the inventories in a given folder: the file name becomes
the name of the object to be evaluated. The function also converts raw data to measurement
units compatible with those of reference prices.
This function can optionally load also critical weights from an Excel file. Such weights are project-specific.
Some materials are considered critical by the European Commission of the International Energy Agency (IEA) but are not considered as individual flows by
SimaPro. For example the SimaPro flow "Coal, hard" contains "coking coal" which is critical. The weights are therefore needed to
compute the Critical-CLCC indicators.
}
