% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{price_eurostat_def}
\alias{price_eurostat_def}
\title{Electricity prices - European Union average}
\format{
A tibble with five variables:
\describe{
\item{year}{year}
\item{cur}{currency (EUR)}
\item{cons_label}{label identifying annual consumption level}
\item{cons_code}{code, code identifying annual consumption level}
\item{price}{price, excluding taxes and levies, euro/MJ, current}
}
}
\source{
\url{https://db.nomics.world/Eurostat/nrg_pc_205}
}
\usage{
price_eurostat_def
}
\description{
A dataset containing electricity prices, to be used in the CLCC indicator calculation
as a proxy for the price of electric energy from renewable sources. Data are
from the Eurostat database (provided by DB.Nomics) and refer to the EU and to the price paid non -domestic users
according to different consumption levels. Data are downloaded through the DB.Nomics API (see data-raw
folder for the script used for downloading the series).
}
\keyword{datasets}
