% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{price_imf_def}
\alias{price_imf_def}
\title{Primary commodity prices from the IMF}
\format{
A tibble with five variables:
\describe{
\item{year}{year}
\item{cur}{currency, (USD)}
\item{source}{data source}
\item{code}{code, IMF database}
\item{price}{price, current usd/kg - usd/m3 according to the commodity considered}
}
}
\source{
\url{https://db.nomics.world/IMF/PCPS}
}
\usage{
price_imf_def
}
\description{
A dataset containing primary commodity prices from the IMF database. Such prices
are used as market prices for some raw materials included in the CLCC indicator.
Prices are downloaded through through the DB.Nomics API.
}
\keyword{datasets}
