% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_prices.R
\name{ref_prices}
\alias{ref_prices}
\title{Compute the reference prices to be used for CLCC calculation}
\usage{
ref_prices(h = 10)
}
\arguments{
\item{h}{A numeric value. Time span (in years) over which calculating average prices.
Default is `10`. Maximum and minimum values admitted are `20` and `5` respectively.}
}
\value{
A tibble containing raw material names, and mean, minimum and maximum
prices of the last ten-year period
}
\description{
This function computes the reference prices to be used for CLCC calculation
and Monte Carlo analysis. Prices are expressed in constant euro. The reference
price is the mean of a ten-year period starting from the reference year, which
is the last year for which data are available (this is automatically calculated).
Such prices are computed for all raw materials present in the *db_comm_prices*
master file.
}
